requirements.txt
streamlit
streamlit-webrtc
opencv-python
numpy
onnxruntime
pyyaml
Pillow
